+++ docs/webpack.config.js
@@ -44,19 +44,7 @@ const createExport = fileName => {
                         test: /\.js$/,
                         exclude: /node_modules/,
                         use: {
-                            loader: 'babel-loader',
-                            options: {
-                                presets: [
-                                    ['@babel/preset-env', {
-                                        modules: false,
-                                        // This adds polyfills when needed. Requires core-js dependency.
-                                        // See https://babeljs.io/docs/en/babel-preset-env#usebuiltins
-                                        // Note that you still need to add custom polyfills if necessary (e.g. whatwg-fetch)
-                                        useBuiltIns: 'usage',
-                                        corejs: 3
-                                    }]
-                                ]
-                            }
+                            loader: 'babel-loader'
                         }
                     }
                 ]