+++ public/UseWorker/RELEASE_NOTES.md
@@ -1,3 +1,8 @@
+### 0.3.0 - Thursday, April 9th, 2020
+* Re-implement public API for more customization and type safety
+* Add auto-serialization for complex message passing
+* Added synchronous fallback for failed workers
+
 ### 0.2.0 - Tuesday, April 7th, 2020
 * Adjust Femto dependencies and switch to webpack for bundling workers